% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{calculate_IS_dependent_shift}
\alias{calculate_IS_dependent_shift}
\title{calculate_IS_dependent_shift}
\usage{
calculate_IS_dependent_shift(sample, set.env = F)
}
\arguments{
\item{sample}{the name of the sample / file batch}

\item{set.env}{boolean if the environment of the R session should be setup from scratch}
}
\description{
Checks internal standards in each sample and calculates potential shifts
}
