% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{errors}
\alias{errors}
\title{errors}
\usage{
errors(e)
}
\description{
function to handle errors from tryCatch functions within the code, especially for parallelization
}
\details{
@param e the error from the tryCatch function
@export
}
