% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{extract_first_two}
\alias{extract_first_two}
\title{extract_first_two}
\usage{
extract_first_two(x)
}
\arguments{
\item{x}{string to extract first two values}
}
\description{
Get the two first values of a string
}
