% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{get_peaks_all}
\alias{get_peaks_all}
\title{get_peaks_all}
\usage{
get_peaks_all(
  sample_path,
  sample,
  RT_range,
  Multiple,
  method_time,
  FullscanMS1,
  Scan_filters,
  precursor_window,
  outer_search_window_multiple_peaks,
  inner_search_window_multiple_peaks,
  outer_intensity_ratio_multiple_peaks,
  inner_intensity_ratio_multiple_peaks,
  References_refined,
  compounds_shift_corrected,
  ppm_val,
  Times_shift_r = 0,
  Times_shift_l = 0,
  use_shifting_search_window = F,
  maximum_nr_of_search_window_extensions,
  alphabet_size,
  zigzag_trigger_threshold,
  normal_background_quantile,
  higher_background_quantile,
  extended_baseline_factor,
  maximum_nr_of_stagnant_intensity_values_peaktop,
  intensity_factor_decrease_peaktop,
  intensity_factor_increase_edges,
  intensity_factor_decrease_edges,
  density_factor_increase_edges,
  density_factor_decrease_edges,
  minimum_background_ratio,
  width_smoothing,
  maximum_nr_of_a,
  minimum_nr_of_high_intensity_letters,
  maximum_peak_width,
  minimum_peak_area,
  minimum_nr_of_datapoints_per_peak,
  minimum_r2_intensity_dependent_shift,
  z,
  Reference
)
}
\arguments{
\item{sample_path}{path to the sample / file batch}

\item{sample}{the name of the sample / file batch}

\item{RT_range}{the retention time range defining the window checked for peaks and patterns in each direction}

\item{Multiple}{file defining peak patterns if multiple peaks occur in the m/z channel within RT_range}

\item{method_time}{the time of the MS measurement}

\item{FullscanMS1}{the scan filter used for analysis of the main peaks}

\item{Scan_filters}{list of scan filters used for screening, mainly used for MS2 annotation}

\item{precursor_window}{the m/z window of precursor masses included when searching for MS2 scans in .mzML files}

\item{outer_search_window_multiple_peaks}{the maximum time in minutes before and after the aimed retention time, defining the search window if peak patterns are expected for a wider RT range}

\item{inner_search_window_multiple_peaks}{the narrow time in minutes before and after the aimed retention time, defining the search window if peak patterns are expected for a narrow RT range}

\item{outer_intensity_ratio_multiple_peaks}{the intensity ratio based on the maximum intensity within the outer RT range, used to filter valid peaks, usually a higher ratio}

\item{inner_intensity_ratio_multiple_peaks}{the intensity ratio based on the maximum intensity within the inner RT range, used to filter valid peaks, usually a lower ratio}

\item{References_refined}{the names of References, sorted by value if possible}

\item{compounds_shift_corrected}{the list of targets with retention times corrected}

\item{ppm_val}{allowed variability of target m/z in ppm}

\item{Times_shift_r}{the number of times the search window was extended to the right}

\item{Times_shift_l}{the number of times the search window was extended to the left}

\item{alphabet_size}{the alphabet size used for SAX functions}

\item{zigzag_trigger_threshold}{defines the maximum fraction of the length of the EIC which can consist of zigzag patterns before triggering a more inclusive background calculation}

\item{normal_background_quantile}{the quantile used to define background intensities from the background EIC}

\item{higher_background_quantile}{the quantile used to define background intensities from the background EIC if the chromatogram is showing quality flags}

\item{extended_baseline_factor}{a ratio applied to the baseline to define values which are considered as baseline}

\item{maximum_nr_of_stagnant_intensity_values_peaktop}{the maximum number of intensities without significant decrease after which the start and end time of the peaktop is defined}

\item{intensity_factor_decrease_peaktop}{the factor applied to the last peak-assigned intensity which needs to be undercut within the maximum number of intensities for a decrease to be considered significant in defining the peaktop}

\item{intensity_factor_increase_edges}{the factor applied to the last peak-assigned intensity which if exceeded triggers a density-dependent counter to assign small increases to the peak while splitting peaks at the peak edges}

\item{intensity_factor_decrease_edges}{the factor applied to the last peak-assigned intensity which needs to be undercut within the density-dependent maximum number of datapoints for a decrease to be considered significant in defining the peak edges}

\item{density_factor_increase_edges}{the factor applied to the intensity density (datapoints within 0.1 minute) which if exceeded by having a constant increase or lack of decrease triggers a stop of the peak detection algorithm and defines the ends of the peak edges}

\item{density_factor_decrease_edges}{the factor applied to the intensity density (datapoints within 0.1 minute) which if exceeded by having too many stagnant intensity values triggers a stop of the peak detection algorithm and defines the ends of the peak edges}

\item{minimum_background_ratio}{the minimum ratio expected for a valid chromatogram of maximum intensity and the normal background quantile}

\item{width_smoothing}{a numeric value needed to define the density of smoothing for the EIC}

\item{maximum_nr_of_a}{within SAX the letter a defines values with the lowest values and a valid peak shall only have a distinct number of these low values to avoid too flat peaks}

\item{minimum_nr_of_high_intensity_letters}{dependent on the selected SAX alphabet, the SAX sequence of a valid peak should have at least the respective number of letters defining high intensities}

\item{maximum_peak_width}{the maximum allowed peak width in minutes}

\item{minimum_peak_area}{the value of a peak area which needs to be exceeded to be a valid peak}

\item{minimum_nr_of_datapoints_per_peak}{the minimal number of data points needed to define a valid peak}

\item{minimum_r2_intensity_dependent_shift}{the lowest accepted coefficient of determination for linear regression used to identify valid intensity-dependent shifts within the calibration}

\item{z}{index of the compound within compounds_shift_corrected to be analyzed}

\item{Reference}{index of the reference file to be used}

\item{is_centroided}{relevant for mzML files if scan mode is centroided ir not}
}
\description{
Uses references to retrieve a list of possible peaks for the respective target
}
