% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{IS_analysis}
\alias{IS_analysis}
\title{IS_analysis}
\usage{
IS_analysis(
  sample_path,
  sample,
  SAX_reference_table,
  Solvent_Blank_ID,
  Multiple,
  RT_range,
  files,
  Peaks,
  Peaks_list,
  compounds_shift_corrected,
  FullscanMS1,
  Scan_filters,
  ppm_val,
  method_time,
  alphabet_size,
  zigzag_trigger_threshold,
  normal_background_quantile,
  higher_background_quantile,
  minimum_background_ratio,
  extended_baseline_factor,
  maximum_nr_of_stagnant_intensity_values_peaktop,
  intensity_factor_decrease_peaktop,
  intensity_factor_increase_edges,
  intensity_factor_decrease_edges,
  density_factor_increase_edges,
  density_factor_decrease_edges,
  width_smoothing,
  width_factor_background,
  sample_search_window,
  maximum_nr_of_a,
  minimum_nr_of_high_intensity_letters,
  maximum_peak_width,
  minimum_peak_area,
  minimum_nr_of_datapoints_per_peak,
  m,
  t,
  j
)
}
\arguments{
\item{sample_path}{path to the sample / file batch}

\item{sample}{name of the sample / file batch}

\item{SAX_reference_table}{the SAX distance matrix used for calculating MINDIST}

\item{Solvent_Blank_ID}{the name ID of solvent blanks with no IS present}

\item{Multiple}{file defining peak patterns if multiple peaks occur in the m/z channel within RT_range}

\item{RT_range}{the retention time range defining the window checked for peaks and patterns in each direction}

\item{files}{names of the raw data files to be analyzed}

\item{Peaks}{results of peak annotation, data frame of peak characteristics from references}

\item{Peaks_list}{Peaks as list per reference}

\item{compounds_shift_corrected}{the list of targets with retention times corrected}

\item{FullscanMS1}{the scan filter used for analysis of the main peaks}

\item{Scan_filters}{list of scan filters used for screening, mainly used for MS2 annotation}

\item{ppm_val}{allowed variability of target m/z in ppm}

\item{method_time}{the time of the MS measurement}

\item{alphabet_size}{the alphabet size used for SAX functions}

\item{zigzag_trigger_threshold}{defines the maximum fraction of the length of the EIC which can consist of zigzag patterns before triggering a more inclusive background calculation}

\item{normal_background_quantile}{the quantile used to define background intensities from the background EIC}

\item{higher_background_quantile}{the quantile used to define background intensities from the background EIC if the chromatogram is showing quality flags}

\item{minimum_background_ratio}{the minimum ratio expected for a valid chromatogram of maximum intensity and the normal background quantile}

\item{extended_baseline_factor}{a ratio applied to the baseline to define values which are considered as baseline}

\item{maximum_nr_of_stagnant_intensity_values_peaktop}{the maximum number of intensities without significant decrease after which the start and end time of the peaktop is defined}

\item{intensity_factor_decrease_peaktop}{the factor applied to the last peak-assigned intensity which needs to be undercut within the maximum number of intensities for a decrease to be considered significant in defining the peaktop}

\item{intensity_factor_increase_edges}{the factor applied to the last peak-assigned intensity which if exceeded triggers a density-dependent counter to assign small increases to the peak while splitting peaks at the peak edges}

\item{intensity_factor_decrease_edges}{the factor applied to the last peak-assigned intensity which needs to be undercut within the density-dependent maximum number of datapoints for a decrease to be considered significant in defining the peak edges}

\item{density_factor_increase_edges}{the factor applied to the intensity density (datapoints within 0.1 minute) which if exceeded by having a constant increase or lack of decrease triggers a stop of the peak detection algorithm and defines the ends of the peak edges}

\item{density_factor_decrease_edges}{the factor applied to the intensity density (datapoints within 0.1 minute) which if exceeded by having too many stagnant intensity values triggers a stop of the peak detection algorithm and defines the ends of the peak edges}

\item{width_smoothing}{a numeric value needed to define the density of smoothing for the EIC}

\item{width_factor_background}{the factor applied to the peak width to define a background window}

\item{sample_search_window}{the search window in minutes added to both ends of the expected retention time used to screen for the peak of interest}

\item{maximum_nr_of_a}{within SAX the letter a defines values with the lowest values and a valid peak shall only have a distinct number of these low values to avoid too flat peaks}

\item{minimum_nr_of_high_intensity_letters}{dependent on the selected SAX alphabet, the SAX sequence of a valid peak should have at least the respective number of letters defining high intensities}

\item{maximum_peak_width}{the maximum allowed peak width in minutes}

\item{minimum_peak_area}{the value of a peak area which needs to be exceeded to be a valid peak}

\item{minimum_nr_of_datapoints_per_peak}{the minimal number of data points needed to define a valid peak}

\item{t}{index of the compound within Peaks}

\item{j}{index of the file to be analyzed in files}

\item{is_centroided}{relevant for mzML files if scan is centroided or not}
}
\description{
Checks for shift of internal standards per sample
}
