% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{letter2number}
\alias{letter2number}
\title{letter to number}
\usage{
letter2number(x)
}
\arguments{
\item{x}{the letter to be transformed}
}
\description{
Transform a letter into a number
}
