% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{Remove_ion}
\alias{Remove_ion}
\title{Remove_ion}
\usage{
Remove_ion(x)
}
\arguments{
\item{x}{string like the compound name including ion information}
}
\description{
Remove additional information like _M+H separated by underscore
}
