% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{sample_analysis}
\alias{sample_analysis}
\title{sample_analysis}
\usage{
sample_analysis(
  sample_path,
  results_path,
  sample,
  files,
  Peaks,
  Peaks_list,
  Solvent_Blank_ID,
  IS_dependent_shift,
  Intensity_dependent_shift,
  Multiple,
  FullscanMS1,
  precursor_window,
  inner_intensity_ratio_multiple_peaks,
  Scan_filters,
  IS_Assignment,
  compounds_shift_corrected,
  RT_range,
  ppm_val,
  method_time,
  IS_deviation,
  gen.plots,
  use.MINDIST,
  SAX_reference_table,
  alphabet_size,
  zigzag_trigger_threshold,
  normal_background_quantile,
  higher_background_quantile,
  minimum_background_ratio,
  extended_baseline_factor,
  width_smoothing,
  width_factor_background,
  sample_search_window,
  maximum_nr_of_a,
  minimum_nr_of_high_intensity_letters,
  maximum_peak_width,
  minimum_peak_area,
  minimum_nr_of_datapoints_per_peak,
  maximum_allowed_shift,
  maximum_allowed_shift_ratio,
  minimum_cutoff_intensity_factor,
  minimum_confirming_peak_height,
  max_MINDIST,
  minimum_datapoints_per_sample_peak,
  minimum_qualitative_threshold,
  t,
  j,
  rep
)
}
\arguments{
\item{sample_path}{path to the sample / file batch}

\item{results_path}{path to the output directory}

\item{sample}{the name of the sample / file batch}

\item{files}{names of the raw data files to be analyzed}

\item{Peaks}{results of peak annotation, data frame of peak characteristics from references}

\item{Peaks_list}{Peaks as list per reference}

\item{IS_dependent_shift}{a data frame linking samples and shifts of the internal standards}

\item{Intensity_dependent_shift}{a data frame listing compounds which show intensity-dependent shifts in calibrations}

\item{Multiple}{file defining peak patterns if multiple peaks occur in the m/z channel within RT_range}

\item{FullscanMS1}{the scan filter used for analysis of the main peaks}

\item{precursor_window}{the m/z window of precursor masses included when searching for MS2 scans in .mzML files}

\item{inner_intensity_ratio_multiple_peaks}{the intensity ratio based on the maximum intensity within the inner RT range, used to filter valid peaks, usually a lower ratio}

\item{Scan_filters}{list of scan filters used for screening, mainly used for MS2 annotation}

\item{IS_Assignment}{data frame linking compounds to their assigned internal standards}

\item{compounds_shift_corrected}{the list of targets with retention times corrected}

\item{ppm_val}{allowed variability of target m/z in ppm}

\item{method_time}{the time of the MS measurement}

\item{gen.plots}{boolean if plots should be generated}

\item{use.MINDIST}{boolean if MINDIST should be applied as peak selection criterion}

\item{SAX_reference_table}{the SAX distance matrix used for calculating MINDIST}

\item{alphabet_size}{the alphabet size used for SAX functions}

\item{zigzag_trigger_threshold}{defines the maximum fraction of the length of the EIC which can consist of zigzag patterns before triggering a more inclusive background calculation}

\item{normal_background_quantile}{the quantile used to define background intensities from the background EIC}

\item{higher_background_quantile}{the quantile used to define background intensities from the background EIC if the chromatogram is showing quality flags}

\item{minimum_background_ratio}{the minimum ratio expected for a valid chromatogram of maximum intensity and the normal background quantile}

\item{extended_baseline_factor}{a ratio applied to the baseline to define values which are considered as baseline}

\item{width_smoothing}{a numeric value needed to define the density of smoothing for the EIC}

\item{width_factor_background}{the factor applied to the peak width to define a background window}

\item{sample_search_window}{the search window in minutes added to both ends of the expected retention time used to screen for the peak of interest}

\item{maximum_nr_of_a}{within SAX the letter a defines values with the lowest values and a valid peak shall only have a distinct number of these low values to avoid too flat peaks}

\item{minimum_nr_of_high_intensity_letters}{dependent on the selected SAX alphabet, the SAX sequence of a valid peak should have at least the respective number of letters defining high intensities}

\item{maximum_peak_width}{the maximum allowed peak width in minutes}

\item{minimum_peak_area}{the value of a peak area which needs to be exceeded to be a valid peak}

\item{minimum_nr_of_datapoints_per_peak}{the minimal number of data points needed to define a valid peak}

\item{maximum_allowed_shift}{the maximum shift allowed in minutes}

\item{maximum_allowed_shift_ratio}{the maximum fraction of IS-dependent shift of the allowed shift, only relevant if allowed shift exceeds maximum value}

\item{minimum_confirming_peak_height}{the minimum height a confirming ion should exceed if expected}

\item{max_MINDIST}{a value used as reference to interpret if the MINDIST of two SAX sequences is significantly different}

\item{minimum_datapoints_per_sample_peak}{the minimal number of data points which should define a valid peak in target screening}

\item{t}{index of the compound within Peaks}

\item{j}{index of the file to be analyzed in files}

\item{rep}{boolean if the function was already called before}

\item{is_centroidedrelevant}{for mzML files if scan is centroided or not}

\item{maximum_cutoff_intensity}{if the intensity of cutoffs which split peaks exceed this value, the cutoff is considered a peak-splitting cutoff and thereby enforced}
}
\description{
Analysis of each sample. Takes shifts and confirming ions into account.
}
