% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{SAX_consensus}
\alias{SAX_consensus}
\title{SAX_consensus}
\usage{
SAX_consensus(vector_of_strings)
}
\arguments{
\item{vector_of_strings}{vector of SAX strings to build consensus}
}
\description{
generate a consensus sequence from SAX strings
}
