% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{SAX_mindist}
\alias{SAX_mindist}
\title{SAX_mindist}
\usage{
SAX_mindist(string1, string2, Nr_of_data_points, SAX_reference_table)
}
\arguments{
\item{string1}{First SAX string}

\item{string2}{Second SAX string}

\item{Nr_of_data_points}{Number of data points which were summarized by the SAX strings}
}
\description{
Uses a SAX distance matrix to calculate the MINDIST between two SAX strings
}
