% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{setup_environment}
\alias{setup_environment}
\title{setup_environment}
\usage{
setup_environment(sample)
}
\arguments{
\item{sample}{the name of the sample / file batch}
}
\description{
sets up the working environment within the R session
}
