% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{sum_letters}
\alias{sum_letters}
\title{sum letters}
\usage{
sum_letters(x, alphabet_size)
}
\arguments{
\item{x}{SAX sequence to analyze}

\item{alphabet_size}{the size of the alphabet selected for SAX}
}
\description{
Counts the occurrence of the highest four letters in the selected alphabet
}
