% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{target_screening_of_files}
\alias{target_screening_of_files}
\title{target_screening_of_files}
\usage{
target_screening_of_files(
  sample,
  set.env = F,
  gen.plots,
  use.MINDIST,
  use.area
)
}
\arguments{
\item{sample}{the name of the sample / file batch}

\item{set.env}{boolean if the environment of the R session should be setup from scratch}

\item{gen.plots}{boolean if plots should be generated}

\item{use.MINDIST}{boolean if MINDIST should be applied as peak selection criterion}

\item{use.area}{boolean, if TRUE uses peak area, otherwise peak height}
}
\description{
Performs the target screening for each file within the sample batch
}
