% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{transformi}
\alias{transformi}
\title{transformi}
\usage{
transformi(y)
}
\arguments{
\item{y}{the array of values to be normalized}
}
\value{
normalized array
}
\description{
Used to derive logarithmic, normalized values for SAX analysis
}
