% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Documentation.R
\name{calc_shift}
\alias{calc_shift}
\title{calc_shift}
\usage{
calc_shift(
  sample_path,
  References,
  IS,
  ppm_val,
  method_time,
  FullscanMS1,
  alphabet_size,
  zigzag_trigger_threshold,
  extended_baseline_factor,
  maximum_nr_of_a,
  minimum_nr_of_high_intensity_letters,
  minimum_peak_area,
  maximum_peak_width,
  i,
  q
)
}
\arguments{
\item{sample_path}{path to the sample / file batch}

\item{References}{vector of names of reference files}

\item{IS}{target list of internal standards}

\item{ppm_val}{allowed variability of target m/z in ppm}

\item{method_time}{the time of the MS measurement}

\item{FullscanMS1}{respective scan filter in MS1}

\item{alphabet_size}{the alphabet size used for SAX functions}

\item{zigzag_trigger_threshold}{defines the maximum fraction of the length of the EIC which can consist of zigzag patterns before triggering a more inclusive background calculation}

\item{extended_baseline_factor}{a ratio applied to the baseline to define values which are considered as baseline}

\item{maximum_nr_of_a}{within SAX the letter a defines values with the lowest values and a valid peak shall only have a distinct number of these low values to avoid too flat peaks}

\item{minimum_nr_of_high_intensity_letters}{dependent on the selected SAX alphabet, the SAX sequence of a valid peak should have at least the respective number of letters defining high intensities}

\item{minimum_peak_area}{the value of a peak area which needs to be exceeded to be a valid peak}

\item{maximum_peak_width}{the maximum allowed peak width in minutes}

\item{i}{index of the standard from IS}

\item{q}{index of the reference from References}
}
\description{
calculates the shift of expected and measured retention time of internal standards in the sample
}
